api/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
passlib
pyjwt
